organizing constitutive models based on types and moving all interfaces from F to grad u so it will be consistent across all physics.
